Core support for position filtering plugins

+++ core/ACMediaNode.h
@@ -42,9 +42,9 @@ using std::vector;
 
 struct ACPoint
 {
-    float x, y, z;
-    ACPoint():x(0.0f),y(0.0f),z(0.0f){}
-    ACPoint(float _x, float _y, float _z):x(_x),y(_y),z(_z){}
+    double x, y, z;
+    ACPoint():x(0.0),y(0.0),z(0.0){}
+    ACPoint(double _x, double _y, double _z):x(_x),y(_y),z(_z){}
 };
 
 class ACMediaNode {
